Refactor CertificateInfoService and kapua-security-certificate-api fixes
