+++ qa/integration/src/test/resources/locator.xml
@@ -46,8 +46,8 @@
 
         <api>org.eclipse.kapua.service.certificate.CertificateService</api>
         <api>org.eclipse.kapua.service.certificate.CertificateFactory</api>
-        <api>org.eclipse.kapua.service.certificate.CertificateInfoService</api>
-        <api>org.eclipse.kapua.service.certificate.CertificateInfoFactory</api>
+        <api>org.eclipse.kapua.service.certificate.info.CertificateInfoService</api>
+        <api>org.eclipse.kapua.service.certificate.info.CertificateInfoFactory</api>
 
         <!-- datastore -->
         <api>org.eclipse.kapua.service.datastore.MessageStoreService</api>
