+++ service/security/certificate/api/src/main/java/org/eclipse/kapua/service/certificate/KeyUsage.java
@@ -11,15 +11,45 @@
  *******************************************************************************/
 package org.eclipse.kapua.service.certificate;
 
+/**
+ * @since 1.0.0
+ */
 public enum KeyUsage {
 
+    /**
+     * @since 1.0.0
+     */
     DIGITAL_SIGNATURE,
+    /**
+     * @since 1.0.0
+     */
     NON_REPUDIATION,
+    /**
+     * @since 1.0.0
+     */
     KEY_ENCIPHERMENT,
+    /**
+     * @since 1.0.0
+     */
     DATA_ENCIPHERMENT,
+    /**
+     * @since 1.0.0
+     */
     KEY_AGREEMENT,
+    /**
+     * @since 1.0.0
+     */
     KEY_CERT_SIGN,
+    /**
+     * @since 1.0.0
+     */
     CRL_SIGN,
+    /**
+     * @since 1.0.0
+     */
     ENCIPHER_ONLY,
+    /**
+     * @since 1.0.0
+     */
     DECIPHER_ONLY,
 }
