+++ service/security/certificate/api/src/main/java/org/eclipse/kapua/service/certificate/KeyUsageSetting.java
@@ -11,18 +11,40 @@
  *******************************************************************************/
 package org.eclipse.kapua.service.certificate;
 
+/**
+ * @since 1.0.0
+ */
+
 public interface KeyUsageSetting {
 
+    /**
+     * @since 1.0.0
+     */
     KeyUsage getKeyUsage();
 
+    /**
+     * @since 1.0.0
+     */
     void setKeyUsage(KeyUsage keyUsage);
 
+    /**
+     * @since 1.0.0
+     */
     boolean getAllowed();
 
+    /**
+     * @since 1.0.0
+     */
     void setAllowed(boolean allowed);
 
+    /**
+     * @since 1.0.0
+     */
     Boolean getKapuaAllowed();
 
+    /**
+     * @since 1.0.0
+     */
     void setKapuaAllowed(Boolean allowed);
 
 }
