+++ service/security/certificate/api/src/main/java/org/eclipse/kapua/service/certificate/info/CertificateInfoFactory.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2018 Eurotech and/or its affiliates and others
+ * Copyright (c) 2018, 2019 Eurotech and/or its affiliates and others
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -9,14 +9,12 @@
  * Contributors:
  *     Eurotech - initial API and implementation
  *******************************************************************************/
-package org.eclipse.kapua.service.certificate;
+package org.eclipse.kapua.service.certificate.info;
 
 import org.eclipse.kapua.model.KapuaEntityFactory;
 
+/**
+ * @since 1.1.0
+ */
 public interface CertificateInfoFactory extends KapuaEntityFactory<CertificateInfo, CertificateInfoCreator, CertificateInfoQuery, CertificateInfoListResult> {
-
-    CertificateUsage newCertificateUsage(String name);
-
-    KeyUsageSetting newKeyUsageSetting(KeyUsage usage, boolean allowed, Boolean kapuaAllowed);
-
 }
