+++ service/security/certificate/api/src/main/java/org/eclipse/kapua/service/certificate/info/CertificateInfoQuery.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2018 Eurotech and/or its affiliates and others
+ * Copyright (c) 2018, 2019 Eurotech and/or its affiliates and others
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -9,18 +9,27 @@
  * Contributors:
  *     Eurotech - initial API and implementation
  *******************************************************************************/
-package org.eclipse.kapua.service.certificate;
+package org.eclipse.kapua.service.certificate.info;
 
-import org.eclipse.kapua.model.query.KapuaListResult;
-import org.eclipse.kapua.service.certificate.xml.CertificateInfoXmlRegistry;
+import org.eclipse.kapua.model.query.KapuaQuery;
+import org.eclipse.kapua.service.certificate.info.xml.CertificateInfoXmlRegistry;
 
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
+import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
 import javax.xml.bind.annotation.XmlType;
 
-@XmlRootElement(name = "certificateListResult")
+/**
+ * @since 1.1.0
+ */
+@XmlRootElement(name = "query")
 @XmlAccessorType(XmlAccessType.PROPERTY)
-@XmlType(factoryClass = CertificateInfoXmlRegistry.class, factoryMethod = "newListResult")
-public interface CertificateInfoListResult extends KapuaListResult<CertificateInfo> {
+@XmlType(factoryClass = CertificateInfoXmlRegistry.class, factoryMethod = "newQuery")
+public interface CertificateInfoQuery extends KapuaQuery<CertificateInfo> {
+
+    @XmlElement(name = "includeInherited")
+    Boolean getIncludeInherited();
+
+    void setIncludeInherited(Boolean includeInherited);
 }
