+++ service/security/certificate/api/src/main/java/org/eclipse/kapua/service/certificate/info/CertificateInfoService.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2017 Eurotech and/or its affiliates and others
+ * Copyright (c) 2018, 2019 Eurotech and/or its affiliates and others
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -9,27 +9,22 @@
  * Contributors:
  *     Eurotech - initial API and implementation
  *******************************************************************************/
-package org.eclipse.kapua.service.certificate;
+package org.eclipse.kapua.service.certificate.info;
 
 import org.eclipse.kapua.KapuaException;
 import org.eclipse.kapua.model.id.KapuaId;
 import org.eclipse.kapua.model.query.KapuaQuery;
 import org.eclipse.kapua.service.KapuaEntityService;
-import org.eclipse.kapua.service.KapuaNamedEntityService;
-import org.eclipse.kapua.service.KapuaUpdatableEntityService;
-import org.eclipse.kapua.service.config.KapuaConfigurableService;
+import org.eclipse.kapua.service.certificate.CertificateUsage;
 
 import java.util.List;
 
-public interface CertificateInfoService extends KapuaEntityService<CertificateInfo, CertificateInfoCreator>,
-        KapuaNamedEntityService<CertificateInfo>,
-        KapuaUpdatableEntityService<CertificateInfo>,
-        KapuaConfigurableService {
-
+/**
+ * @since 1.1.0
+ */
+public interface CertificateInfoService extends KapuaEntityService<CertificateInfo, CertificateInfoCreator> {
     @Override
     CertificateInfoListResult query(KapuaQuery<CertificateInfo> query) throws KapuaException;
-
-    CertificateInfo generate(CertificateGenerator generator) throws KapuaException;
 
     List<CertificateInfo> findAncestorsCertificates(KapuaId scopeId, CertificateUsage usage) throws KapuaException;
 }
