+++ service/security/certificate/api/src/main/java/org/eclipse/kapua/service/certificate/xml/CertificateXmlRegistry.java
@@ -12,12 +12,13 @@
 package org.eclipse.kapua.service.certificate.xml;
 
 import org.eclipse.kapua.locator.KapuaLocator;
-import org.eclipse.kapua.service.certificate.CertificateGenerator;
 import org.eclipse.kapua.service.certificate.Certificate;
 import org.eclipse.kapua.service.certificate.CertificateCreator;
 import org.eclipse.kapua.service.certificate.CertificateFactory;
+import org.eclipse.kapua.service.certificate.CertificateGenerator;
 import org.eclipse.kapua.service.certificate.CertificateListResult;
 import org.eclipse.kapua.service.certificate.CertificateQuery;
+import org.eclipse.kapua.service.certificate.CertificateUsage;
 
 import javax.xml.bind.annotation.XmlRegistry;
 
@@ -45,5 +46,9 @@ public class CertificateXmlRegistry {
 
     public CertificateGenerator newCertificateGenerator() {
         return FACTORY.newCertificateGenerator();
+    }
+
+    public CertificateUsage newCertificateUsage() {
+        return FACTORY.newCertificateUsage(null);
     }
 }
